[SofaKernel] move Vec (de)serialization operators into its mother class fixed_array before these operators were duplicated in both classes !
